requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torchvision>=0.15.0
flwr>=1.4.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
transformers>=4.20.0
Pillow>=8.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
jupyter>=1.0.0
